Enable both RSA + ECDSA keys to be written to SSM
This commit adds some default variables and locals for naming of SSM
params.

We expect SSM params to be written in the form `/foo/bar` or
`/foo/bar_badger` where foo is var.ssm_path_prefix and bar is passed in
explicitly, or else we default to label.id where we replace the
delimiter with underscores to maintain consistency.

Both RSA + ECDSA key algo’s are supported.